moved tests too see what travis/codecov does
